Cleanup ViewModel, move logic to StopsResulsMAnager
