+++ feature/trip-planner/ui/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/di/ViewModelModule.kt
@@ -11,6 +11,8 @@ import xyz.ksharma.krail.trip.planner.ui.settings.SettingsViewModel
 import xyz.ksharma.krail.trip.planner.ui.timetable.TimeTableViewModel
 import xyz.ksharma.krail.trip.planner.ui.alerts.ServiceAlertsViewModel
 import xyz.ksharma.krail.trip.planner.ui.datetimeselector.DateTimeSelectorViewModel
+import xyz.ksharma.krail.trip.planner.ui.searchstop.RealStopResultsManager
+import xyz.ksharma.krail.trip.planner.ui.searchstop.StopResultsManager
 import xyz.ksharma.krail.trip.planner.ui.themeselection.ThemeSelectionViewModel
 
 val viewModelsModule = module {
@@ -44,4 +46,6 @@ val viewModelsModule = module {
             ioDispatcher = get(named(IODispatcher)),
         )
     }
+
+    single<StopResultsManager> { RealStopResultsManager(get(), get(), get()) }
 }
