+++ feature/trip-planner/ui/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/searchstop/SearchStopViewModel.kt
@@ -17,24 +17,12 @@ import xyz.ksharma.krail.core.analytics.Analytics
 import xyz.ksharma.krail.core.analytics.AnalyticsScreen
 import xyz.ksharma.krail.core.analytics.event.AnalyticsEvent
 import xyz.ksharma.krail.core.analytics.event.trackScreenViewEvent
-import xyz.ksharma.krail.core.log.log
-import xyz.ksharma.krail.core.remote_config.flag.Flag
-import xyz.ksharma.krail.core.remote_config.flag.FlagKeys
-import xyz.ksharma.krail.core.remote_config.flag.asBoolean
-import xyz.ksharma.krail.core.remote_config.flag.toStopsIdList
-import xyz.ksharma.krail.sandook.Sandook
-import xyz.ksharma.krail.sandook.SelectProductClassesForStop
-import xyz.ksharma.krail.trip.planner.network.api.service.TripPlanningService
-import xyz.ksharma.krail.trip.planner.ui.searchstop.StopResultMapper.toStopResults
-import xyz.ksharma.krail.trip.planner.ui.state.TransportMode
-import xyz.ksharma.krail.trip.planner.ui.state.TransportModeSortOrder
 import xyz.ksharma.krail.trip.planner.ui.state.searchstop.SearchStopState
 import xyz.ksharma.krail.trip.planner.ui.state.searchstop.SearchStopUiEvent
 
 class SearchStopViewModel(
-    private val tripPlanningService: TripPlanningService,
     private val analytics: Analytics,
-    private val sandook: Sandook,
+    private val stopResultsManager: StopResultsManager,
 ) : ViewModel() {
 
     private val _uiState: MutableStateFlow<SearchStopState> = MutableStateFlow(SearchStopState())
@@ -44,14 +32,6 @@ class SearchStopViewModel(
         }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), SearchStopState())
 
     private var searchJob: Job? = null
-
-    private val isLocalStopsEnabled: Boolean by lazy {
-        false
-    }
-    private val highlightStopIdList: List<String> by lazy {
-        emptyList()
-//        flag.getFlagValue(FlagKeys.HIGH_PRIORITY_STOP_IDS.key).toStopsIdList()
-    }
 
     fun onEvent(event: SearchStopUiEvent) {
         when (event) {
@@ -70,58 +50,13 @@ class SearchStopViewModel(
         searchJob = viewModelScope.launch {
             delay(150)
             runCatching {
-                val stopResults = fetchStopResults(query)
+                val stopResults = stopResultsManager.fetchStopResults(query)
                 updateUiState { displayData(stopResults) }
             }.getOrElse {
                 delay(1500) // buffer for API response before displaying error.
                 updateUiState { displayError() }
             }
         }
-    }
-
-    private suspend fun fetchStopResults(query: String): List<SearchStopState.StopResult> =
-        if (isLocalStopsEnabled) {
-            log("fetchStopResults from LOCAL_STOPS")
-            val resultsDb: List<SelectProductClassesForStop> =
-                sandook.selectStops(stopName = query, excludeProductClassList = listOf())
-
-            val results = resultsDb
-                .map { it.toStopResult() }
-                .let {
-                    filterProductClasses(
-                        stopResults = it,
-                        excludedProductClasses = listOf(TransportMode.Ferry().productClass).toImmutableList()
-                    )
-                }
-                .let(::prioritiseStops)
-                .take(50)
-
-            results
-        } else {
-            log("fetchStopResults from REMOTE")
-            val response = tripPlanningService.stopFinder(stopSearchQuery = query)
-            log("response VM: $response")
-            response.toStopResults()
-        }
-
-    // TODO - move to another file and add UT for it. Inject and use.
-    private fun prioritiseStops(stopResults: List<SearchStopState.StopResult>): List<SearchStopState.StopResult> {
-        val sortedTransportModes = TransportMode.sortedValues(TransportModeSortOrder.PRIORITY)
-        val transportModePriorityMap = sortedTransportModes.mapIndexed { index, transportMode ->
-            transportMode.productClass to index
-        }.toMap()
-
-        return stopResults.sortedWith(compareBy(
-            { stopResult ->
-                if (stopResult.stopId in highlightStopIdList) 0 else 1
-            },
-            { stopResult ->
-                stopResult.transportModeType.minOfOrNull {
-                    transportModePriorityMap[it.productClass] ?: Int.MAX_VALUE
-                } ?: Int.MAX_VALUE
-            },
-            { it.stopName }
-        ))
     }
 
     private fun SearchStopState.displayData(stopsResult: List<SearchStopState.StopResult>) = copy(
@@ -153,12 +88,3 @@ fun filterProductClasses(
         productClasses.any { it !in excludedProductClasses }
     }
 }
-
-/// TODO - move to mapper:
-private fun SelectProductClassesForStop.toStopResult() = SearchStopState.StopResult(
-    stopId = stopId,
-    stopName = stopName,
-    transportModeType = this.productClasses.split(",").mapNotNull {
-        TransportMode.toTransportModeType(it.toInt())
-    }.toImmutableList(),
-)
